+++ package.js
@@ -41,7 +41,6 @@ function configure( api ){
     api.use( 'pwix:modal@2.0.0' );
     api.use( 'pwix:notes@1.0.0-rc' );
     api.use( 'pwix:plus-button@1.0.0-rc' );
-    api.use( 'pwix:roles@1.3.0' );
     api.use( 'pwix:tabbed@1.0.0-rc' );
     api.use( 'pwix:tabular@1.0.0-rc' );
     api.use( 'pwix:typed-message@1.2.0' );